move integration test to a more appropriate home
