--- integration_tests/test_asset_tracking.py
@@ -1,39 +1,0 @@
-# -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
-#
-# Copyright (C) 2017 Canonical Ltd
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as
-# published by the Free Software Foundation.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import os
-import yaml
-
-from testtools.matchers import FileExists
-
-import integration_tests
-
-
-class AssetTrackingTestCase(integration_tests.TestCase):
-
-    def test_pull(self):
-        project_dir = 'asset-tracking'
-        self.run_snapcraft('pull', project_dir)
-
-        state_file = os.path.join(
-            self.parts_dir, 'asset-tracking', 'state', 'pull')
-        self.assertThat(state_file, FileExists())
-        with open(state_file) as f:
-            state = yaml.load(f)
-
-        # Verify that the correct version of 'hello' is installed
-        self.assertTrue(len(state.assets['stage-packages']) > 0)
-        self.assertIn('hello=2.10-1', state.assets['stage-packages'])
